Video update: - defined codec IDs for H263-1998 & H263-2000 (AFAIK the payloads are identical, they differ only by the supported MIME attributes and H263 annexes) - video codec runtime selection - video scaling support in MpvoGdi (rendered video is scaled to match the client area of output window, it's the client's responsibility to set the desired size) - various bugfixes & tweaks for H263
Regarding the codec negotiation: there is still a lot of work to be done in order to be able to request specific bitrate, framerate, video size etc. Currently I commented out code which was dealing with obsolete MIME attributes like "size" or "imagesize", however I'm thinking of generalizing this process slightly.

Thanks to Andrzej Ciarkowski for this patch.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi-media-update@10278 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/contrib/VideoSupport/VideoCapture.cpp
@@ -1594,9 +1594,9 @@ void VideoCapture::Impl::BestFormatMatch(const AM_MEDIA_TYPE& type, const VIDEO_
 
 
 
-	if (format.GetWidth() <= caps.MinOutputSize.cx && 
-		(caps.MinOutputSize.cx < nextW || 
-			(caps.MinOutputSize.cx == nextW && 
+	if (format.GetWidth() <= size_t(caps.MinOutputSize.cx) && 
+		(size_t(caps.MinOutputSize.cx) < nextW || 
+			(size_t(caps.MinOutputSize.cx) == nextW && 
 			GetVideoSurfaceDistance(format.GetSurface(), surface) < GetVideoSurfaceDistance(format.GetSurface(), nextSurface))))
 	{
 		nextW = caps.MinOutputSize.cx;
@@ -1604,9 +1604,9 @@ void VideoCapture::Impl::BestFormatMatch(const AM_MEDIA_TYPE& type, const VIDEO_
 		nextSurface = surface;
 	}
 
-	if (format.GetWidth() >= caps.MaxOutputSize.cx && 
-		(caps.MaxOutputSize.cx > prevW || 
-			(caps.MaxOutputSize.cx == prevW && 
+	if (format.GetWidth() >= size_t(caps.MaxOutputSize.cx) && 
+		(size_t(caps.MaxOutputSize.cx) > prevW || 
+			(size_t(caps.MaxOutputSize.cx) == prevW && 
 			GetVideoSurfaceDistance(format.GetSurface(), surface) < GetVideoSurfaceDistance(format.GetSurface(), prevSurface))))
 	{
 		prevW = caps.MaxOutputSize.cx;
